ai-methodologist/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mattermostdriver==7.3.2
requests==2.32.3
anthropic==0.34.2
python-docx==1.1.2
pdfplumber==0.11.4
python-pptx==1.0.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-mock==3.14.0
